+++ spec/models/question_spec.rb
@@ -3,24 +3,24 @@ require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
 describe Question, "when creating a new question" do
   before(:each) do
     @ss = mock_model(SurveySection)
-    @question = Question.new(:text => "What is your favorite color?", :survey_section => @ss, :is_mandatory => true, :display_order => 1)
+    @question = Question.new(:text => "What is your favorite color?", :survey_section => @ss, :is_mandatory => false, :display_order => 1)
   end
 
   it "should be invalid without text" do
     @question.text = nil
     @question.should have(1).error_on(:text)
   end
-    
+
   it "should have a parent survey section" do
     # this causes issues with building and saving
     # @question.survey_section = nil
     # @question.should have(1).error_on(:survey_section_id)
   end
-    
-  it "should be mandatory by default" do
-    @question.mandatory?.should be_true
+
+  it "should not be mandatory by default" do
+    @question.mandatory?.should be_false
   end
-  
+
   it "should convert pick attribute to string" do
     @question.pick.should == "none"
     @question.pick = :one
@@ -28,7 +28,7 @@ describe Question, "when creating a new question" do
     @question.pick = nil
     @question.pick.should == nil
   end
-  
+
   it "should split the text" do
     @question.split_text.should == "What is your favorite color?"
     @question.split_text(:pre).should == "What is your favorite color?"
@@ -38,11 +38,11 @@ describe Question, "when creating a new question" do
     @question.split_text(:pre).should == "before"
     @question.split_text(:post).should == "after|extra"
   end
-  
+
   it "should have an api_id" do
     @question.api_id.length.should == 36
   end
-  
+
   it "should protect api_id, timestamps" do
     saved_attrs = @question.attributes
     if defined? ActiveModel::MassAssignmentSecurity::Error
@@ -63,11 +63,11 @@ describe Question, "that has answers" do
     Factory(:answer, :question => @question, :display_order => 1, :text => "red")
     Factory(:answer, :question => @question, :display_order => 2, :text => "green")
   end
-  
+
   it "should have answers" do
     @question.answers.should have(3).answers
   end
-  
+
   it "should retrieve those answers in display_order" do
     @question.answers.map(&:display_order).should == [1,2,3]
   end
@@ -79,7 +79,7 @@ describe Question, "that has answers" do
 end
 
 describe Question, "when interacting with an instance" do
-  
+
   before(:each) do
     @question = Factory(:question)
   end
@@ -88,7 +88,7 @@ describe Question, "when interacting with an instance" do
     @question.display_type = nil
     @question.renderer.should == :default
   end
-  
+
   it "should let you know if it is part of a group" do
     @question.question_group = Factory(:question_group)
     @question.solo?.should be_false
@@ -97,7 +97,7 @@ describe Question, "when interacting with an instance" do
     @question.solo?.should be_true
     @question.part_of_group?.should be_false
   end
-  
+
   require 'mustache'
   class FakeMustacheContext < ::Mustache
     def site
@@ -111,7 +111,7 @@ describe Question, "when interacting with an instance" do
       @question.text = "You are in {{site}}"
       @question.render_question_text(FakeMustacheContext).should == "You are in Northwestern"
   end
-        
+
 end
 
 describe Question, "with dependencies" do
@@ -131,5 +131,5 @@ describe Question, "with dependencies" do
     @question.destroy
     Dependency.find_by_id(dep_id).should be_nil
   end
-  
+
 end